Implement basic audio effect

+++ automotive/src/main/java/com/example/mediaplayerapp/AudioEqualizer.kt
@@ -15,7 +15,7 @@ class AudioEqualizer {
      * @param bassStrength A intensidade do efeito de reforço de graves.
      * @return Retorna o áudio processado como um array de amostras.
      */
-    external fun applyBassBoostEffect(audioData: ShortArray, bassStrength: Float): ShortArray
+    external fun applyBassBoostEffect(audioData: ByteArray, bassStrength: Float): ByteArray
 
     /**
      * Aplica o efeito de reforço de médios no áudio.
@@ -24,5 +24,5 @@ class AudioEqualizer {
      * @param midrangeStrength A intensidade do efeito de reforço de médios.
      * @return Retorna o áudio processado como um array de amostras.
      */
-    external fun applyMidrangeBoostEffect(audioData: ShortArray, midrangeStrength: Float): ShortArray
+    external fun applyMidrangeBoostEffect(audioData: ByteArray, midrangeStrength: Float): ByteArray
 }
